Refactor updateGameState to use float deltaTime for improved precision and update NetworkConfig InterpolationPeriod for better timing accuracy

+++ SOS/include/network/EmbeddedServer.h
@@ -71,7 +71,7 @@ private:
 
     // Game logic methods
     void createInitialGameObjects();
-    void updateGameState(uint64_t deltaTime);
+    void updateGameState(float deltaTime);
     void detectAndResolveCollisions();
     void sendGameStateToClients();
